feat: refactor gen_begin_tx_ops and gen_end_tx_ops
